requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.10.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
